requirements.txt
psycopg2
pandas
requests
python-dotenv